Commit -  cleaning

+++ src/main/java/com/GoGym/service/TrainerClientService.java
@@ -1,33 +1,27 @@
 package com.GoGym.service;
-
 import com.GoGym.module.TrainerClient;
-import com.GoGym.module.TrainerDetails;
-import com.GoGym.module.TrainerExperience;
 import com.GoGym.module.User;
 import com.GoGym.repository.TrainerClientRepository;
-import com.GoGym.repository.TrainerDetailsRepository;
-import com.GoGym.repository.TrainerExperienceRepository;
 import com.GoGym.repository.UserRepository;
 import jakarta.transaction.Transactional;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-
+import org.springframework.web.server.ResponseStatusException;
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@AllArgsConstructor
 public class TrainerClientService {
 
-    private final TrainerClientRepository trainerClientRepository;
-    private final UserRepository userRepository;
-    private final NotificationService notificationService;
-
     @Autowired
-    public TrainerClientService(TrainerClientRepository trainerClientRepository, UserRepository userRepository, NotificationService notificationService) {
-        this.trainerClientRepository = trainerClientRepository;
-        this.userRepository = userRepository;
-        this.notificationService = notificationService;
-    }
+    private final TrainerClientRepository trainerClientRepository;
+    @Autowired
+    private final UserRepository userRepository;
+    @Autowired
+    private final NotificationService notificationService;
 
     public TrainerClient createTrainerClient(Long trainerId, Long clientId) {
         TrainerClient trainerClient = new TrainerClient();
@@ -50,19 +44,14 @@ public class TrainerClientService {
 
     @Transactional
     public void removeTrainerClient(Long trainerId, Long clientId) {
-        TrainerClient trainerClient = trainerClientRepository.findByTrainer_IdUserAndClient_IdUser(trainerId, clientId)
-                .orElseThrow(() -> new IllegalArgumentException("Nie znaleziono relacji trener-klient"));
-
+        Optional<TrainerClient> trainerClientOpt = trainerClientRepository.findByTrainer_IdUserAndClient_IdUser(trainerId, clientId);
+        if (trainerClientOpt.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Relacja trener-klient nie istnieje");
+        }
+        TrainerClient trainerClient = trainerClientOpt.get();
         User trainer = trainerClient.getTrainer();
         User client = trainerClient.getClient();
-
-        // Usunięcie relacji
         trainerClientRepository.delete(trainerClient);
-
-        // Dodanie powiadomienia dla klienta
         notificationService.createNotification(client, trainer, "trainer_resigned");
     }
-
-
-
 }
